# Cache System Improvements & Performance Fixes