… runelite-client/src/main/java/net/runelite/client/plugins/microbot/util/cache/Rs2ObjectCache.java
… runelite-client/src/main/java/net/runelite/client/plugins/microbot/util/cache/Rs2QuestCache.java
… runelite-client/src/main/java/net/runelite/client/plugins/microbot/util/cache/serialization/CacheSerializationManager.java
… runelite-client/src/main/java/net/runelite/client/plugins/microbot/util/cache/strategy/entity/ObjectUpdateStrategy.java
… runelite-client/src/main/java/net/runelite/client/plugins/microbot/util/player/Rs2Player.java