## Overview

This PR improves the cache system with performance optimizations, asynchronous quest state management, and better integration with existing utilities. The main goal was to fix performance bottlenecks and improve experience through faster bank location detection and more efficient cache operations.

## Key Improvements

### Major Performance Gains

#### Bank Location Detection Speed Improvement (90% faster)
- **Before**: Accessible banks filtering in `getPathAndBankToNearestBank()` took 2500ms+ due to direct client calls for membership status, quest state retrieval and skill requirement retrieval
- **After**: Now takes 0.4ms - 1ms using fully cached data (updated missing `Rs2Player.isMember()` to use cached data)
- **Impact**: Significant improvement in `getNearestBank()` performance when filtering all bank locations

#### Quest Cache Asynchronous Architecture
- **Problem**: Quest state loading was causing problems, assertions, freezing client thread (through long script execution times)
- **Solution**: Complete refactor to use `invokeLater()` for all quest state operations
- **Benefits**: 
  - Eliminates thread blocking during quest state retrieval
  - Prevents deadlocks in event-driven scenarios, multiple events want to update the quest state, now tracking correctly if update is in progress
  - Maintains UI responsiveness during cache updates

### Technical Improvements

## Rs2QuestCache Asynchronous Refactoring

### Why We Now Use Asynchronous Updates with `invokeLater()`

The quest cache system was completely refactored to use asynchronous operations for all quest state loading. This was necessary because of script threading issues when running on client thread and updating quest state on client thread, during events:

#### Thread Context Problems
- Scripts often request quest states from background threads and simultaneously from events
- Quest-related events like `VarbitChanged` and `ChatMessage` execute on the client thread
- Direct client calls from event handlers could cause problems when executing the `client.runScript(ScriptID.QUEST_STATUS_GET, id);` within the `enum Quest`

#### The `invokeLater()` Solution

```java
// Before: Direct blocking calls, waiting for the script to finish
QuestState state = quest.getState(Microbot.getClient());

// After: Asynchronous with callback
loadQuestStateFromClientAsync(quest, state -> {
    // Handle state asynchronously
    updateQuestState(quest, state);
});
```

#### Key Benefits
- All quest loading happens via `invokeLater()` for consistency
- Methods return immediately, preventing UI freezes
- Quest granular updating and tracking for each specific quest, coordination prevents duplicate loading requests
- Event handlers can safely trigger quest updates without blocking

#### Implementation Details
- Helper methods provide clean separation: `executeQuestStateLoad()`, `handleLoadCompletion()`, `safeCallback()`
- Centralized error handling with graceful fallbacks
- Prevents duplicate updates for the same quest using `questUpdatesInProgress`
- Automatic cleanup of stale update entries after 10 seconds

### Cache Loading Improvements

#### Retry Logic for Player Validation

```java
private static void retryLoadCacheWithValidation(String newRsProfileKey, int attemptCount) {
    // Validates player is fully loaded before cache operations
    // Retries up to 10 times with 1-second delays
    // Prevents cache loading failures during login transitions
}
```

#### Improved Cache Deserialization

- Only loads entries that aren't already present in cache (newer data takes precedence)
- Reduces unnecessary cache overwrites during profile switches
- Improved logging for tracking skipped vs loaded entries

### Profile-Based Cache Persistence System

#### Intelligent Cache Loading and Saving on Profile Switches

The cache system now intelligently manages persistence based on RuneScape profile changes:

```java
@Subscribe
public void onRuneScapeProfileChanged(RuneScapeProfileChanged event) {
    String newProfile = event.getNewProfile();
    String oldProfile = event.getPreviousProfile();
    if ((newProfile != null && !newProfile.isEmpty()) &&
        (oldProfile == null || oldProfile.isEmpty() || !newProfile.equals(oldProfile))) {
        log.info("Received RuneScape profile change event from '{}' to '{}'", oldProfile, newProfile);
        
        // Save caches for the old profile before switching
        Rs2CacheManager.savePersistentCaches(oldProfile);
        Rs2CacheManager.setUnknownInitialCacheState();
        
        // Load caches for the new profile
        Rs2CacheManager.loadCacheStateFromConfig(newProfile);
    }
}
```

#### Benefits of Profile-Based Cache Management

- **Isolated Cache Data**: Each RuneScape profile maintains its own cache data
- **Automatic Persistence**: Caches are saved when switching profiles and on client shutdown
- **Memory Efficiency**: Only loads cache data for the current active profile
- **Data Integrity**: Prevents cache data contamination between different characters



#### Scene Scanning for Cache Population

- **Performance-Aware Scanning**: Only performs scene scans when cache is empty or explicitly requested
- **Minimum Scan Interval**: Prevents excessive scanning with 2-second cooldown periods
- **Fallback Scene Scanning**: Automatically scans scene when cached objects are not found

```java
public void performSceneScan(CacheOperations<String, Rs2ObjectModel> cache, boolean force) {
    // Respects minimum scan interval unless forced
    if (!force && (currentTime - lastSceneScan) < MIN_SCAN_INTERVAL_MS) {
        return;
    }
    
    // Only scan if really needed or forced
    if (!force && !needsSceneScan && cache.size() > 10) {
        return;
    }
    // Comprehensive scene scanning logic...
}
```

#### Region Change Detection and Cache Invalidation

The system now detects region changes and properly handles cache invalidation:

- **Automatic Region Tracking**: Monitors current regions vs last known regions
- **Stale Object Cleanup**: Clears cache when regions change (since RuneLite doesn't fire despawn events)
- **Event-Driven Updates**: Uses proper EventBus integration for all object spawn/despawn events

### Rs2ObjectModel Lazy Loading Optimizations

#### Cached Object Composition for Performance

The Rs2ObjectModel now implements lazy loading for object composition data:

```java
public ObjectComposition getObjectComposition() {
    if (this.objectComposition == null) {
        // Lazy load composition data only when needed
        this.objectComposition = Microbot.getClient().getObjectDefinition(getId());
    }
    return this.objectComposition;
}
```

#### Enhanced Object Name Resolution

- **Optimized getName() Method**: Removed unnecessary client thread calls and improved performance
- **Direct Composition Access**: Uses cached object composition for name resolution instead of repeated client calls
- **Simplified Creation Time Tracking**: Eliminated redundant client thread operations during object creation

### MicrobotPlugin Profile-Based Cache Management

#### Intelligent Cache Loading on Profile Switches

The MicrobotPlugin now handles all cache persistence during RuneScape profile changes:

```java
@Subscribe
public void onRuneScapeProfileChanged(RuneScapeProfileChanged event) {
    String newProfile = event.getNewProfile();
    // Save current profile's cache data before switching
    Rs2CacheManager.savePersistentCaches(oldProfile);
    Rs2CacheManager.setUnknownInitialCacheState();
    Rs2CacheManager.loadCacheStateFromConfig(newProfile);
}
```

#### Profile-Based Cache Persistence Benefits

- **Automatic Cache Persistence**: Saves cache data when switching between profiles
- **Isolated Profile Data**: Each RuneScape profile maintains separate cache state
- **Reduced Login Loading**: Cache loading moved to profile change events instead of every login
- **Memory Efficiency**: Only loads cache data for the active profile

### ObjectUpdateStrategy Thread Safety Improvements

#### Asynchronous Event Handling

The ObjectUpdateStrategy now uses `invokeLater()` for all event processing:

```java
@Override
public void handleEvent(final Object event, final CacheOperations<String, Rs2ObjectModel> cache) {
    Microbot.getClientThread().invokeLater(() -> {
        try {
            // All event processing now happens asynchronously
            if (event instanceof GameObjectSpawned) {
                handleGameObjectSpawned((GameObjectSpawned) event, cache);
            }
            // ... other event handlers
        } catch (Exception e) {
            log.error("Error handling event: {}", event.getClass().getSimpleName(), e);
        }
    });
}
```

#### Static Method Optimizations

- **Eliminated Instance Creation**: Made helper methods static to avoid unnecessary object instantiation
- **Simplified Cache Key Generation**: Removed redundant strategy object creation during key generation
- **Enhanced Error Handling**: Added comprehensive exception handling for all event processing

### Rs2CacheManager Robustness Improvements

#### Event Registration Tracking

Added thread-safe event registration tracking to prevent duplicate registrations:

```java
private static AtomicBoolean isEventRegistered = new AtomicBoolean(false);

private static void registerEventHandlers() {
    if (eventBus == null || isEventRegistered.get()) {
        return; // Prevent duplicate registrations
    }
    // ... registration logic
    isEventRegistered.set(true);
}
```

#### Retry Logic for Cache Loading

Implemented robust retry mechanism for cache loading with player validation:

```java
private static void retryLoadCacheWithValidation(String newRsProfileKey, int attemptCount) {
    // Validates player is fully loaded before cache operations
    // Retries up to MAX_CACHE_LOAD_ATTEMPTS with configurable delays
    // Prevents cache loading failures during login transitions
}
```

#### Cache Loading Benefits

- **Player Validation**: Ensures player data is available before loading caches
- **Configurable Retry Logic**: Up to 10 attempts with 1-second delays between retries
- **Thread-Safe Loading**: Prevents duplicate cache loading requests using atomic flags
- **Graceful Failure Handling**: Proper cleanup on loading failures

### Rs2Player Integration with Cache System

#### Membership Status retrieval now uses the cache system

```java
// Before: Direct client calls every time
public static boolean isMember() {
    return Microbot.getClientThread().runOnClientThreadOptional(() ->
            Microbot.getClient().getVarpValue(VarPlayer.MEMBERSHIP_DAYS) > 0
    ).orElse(false);
}

// After: Cache-based lookup
public static boolean isMember() {
    return Rs2VarPlayerCache.getVarPlayerValue(VarPlayerID.ACCOUNT_CREDIT) > 0;
}
```

#### Performance Impact on `getNearestBank()`

The membership check optimization has a significant impact on bank location detection in the `getPathAndBankToNearestBank()` function:

1. `getPathAndBankToNearestBank()` filters all bank locations using `BankLocation::hasRequirements`
2. Each `BankLocation.hasRequirements()` calls `Rs2Player.isMember()` to check membership access
3. Before: Every bank location made a direct client call to check membership status (2500ms+ total, without any caching of the quest state, skill, varbit, varplayer, before the caching system)
4. After: Uses cached values, making the entire filtering operation complete in 0.4-1ms
5. Result: The accessible banks filtering went from 2500ms to under 1ms - a 99.96% performance improvement

This is where the major performance gain comes from. In `getPathAndBankToNearestBank()`, we have this section:

```java
// Measure accessible banks filtering performance, now down to 0.4ms to 1ms, proper usage of the cache in hasRequirements
long accessibleBanksStart = System.nanoTime();
Set<BankLocation> accessibleBanks = allBanks.stream()
        .filter(BankLocation::hasRequirements)
        .collect(Collectors.toSet());
long accessibleBanksTime = System.nanoTime() - accessibleBanksStart;
```

Before the cache integration, this filtering operation would take 2500ms+ because every `BankLocation.hasRequirements()` call triggered direct client calls. Now it uses the cached status, making the entire operation nearly instantaneous.

#### Additional Cache Integration

- Skill level checks in potion effects now use cached skill data
- Prayer status checks use cached values
- Varbit access uses `Rs2VarbitCache` throughout Rs2Player


### Configuration & Persistence

#### Improved Cache Loading Strategy

- Caches are loaded per RuneScape profile
- Player validation before cache operations
- Robust retry mechanism for initialization failures
- Bank cache loading moved to separate thread

#### Serialization Enhancements

- Only loads cache entries that aren't already present
- Better logging for cache load/save operations
- Graceful handling of serialization failures

## Performance Measurements

### Before vs After

| Operation | Before | After | Improvement |
|-----------|--------|-------|-------------|
| Accessible banks filtering in `getPathAndBankToNearestBank()` | 2500ms before any caching | 0.4-3ms | 99.96% faster |
| Quest state loading | Blocking calls | Asynchronous | I have no more UI freezes or assertions, please report if there are still issues |

The performance improvements are most noticeable in:

- Scripts that frequently check bank accessibility
- Banking operations that need to find the nearest bank
- Quest-related plugins that monitor quest progression
- world hopping and login times shouldd have been impoved

## Architecture Changes

### Cache System Integration

- All caches now follow consistent patterns
- Proper event handler registration and cleanup
- Improved startup/shutdown coordination
- Better cache cleanup and resource management

### Thread Safety Improvements

- Quest operations designed for asynchronous execution
- Coordination mechanisms for concurrent access
- All event handlers use deferred execution patterns
- No blocking calls in event context

## Improvements

- Significant speed improvement in bank-related operations
- Quest states load consistently without blocking
- More robust error recovery in cache operations
- Should fix the freezes during cache operations
- Bank location detection is nearly instantaneous
- Consistent performance across different scenarios

## Files Changed

### Core Cache Infrastructure

- `Rs2Cache.java`: Enhanced base cache with better logging
- `Rs2CacheManager.java`: Improved initialization and retry logic
- `CacheSerializationManager.java`: Smart loading with duplicate prevention

### Quest System

- `Rs2QuestCache.java`: Complete async refactor with helper methods
- Quest Loading: All operations now use `invokeLater()` pattern

### Player Integration

- `Rs2Player.java`: Integration with cache system for membership and stats
- Bank Integration: `Rs2Bank.java` performance logging improvements

### Plugin Integration

- `MicrobotPlugin.java`: Updated cache initialization timing
- `Rs2BankData.java`: Restored lazy loading for better performance


## Acknowledgments

Special thanks to @Kryox and @Bizza, and others on the Discord for firguring the problems. Thanks for providing a "quick" solution for the all,i rely appricate your works and feedback! These gives me the opportunity to implement these performance improvements

This caching system overhaul represents a significant step forward in making Microbot scripts more responsive and efficient.

---


<img width="783" height="234" alt="Screenshot from 2025-07-29 11-34-03" src="https://github.com/user-attachments/assets/3f3260b6-d4a2-4efc-80cf-1abd419632e8" />
<img width="500" height="126" alt="Screenshot from 2025-07-29 11-33-52" src="https://github.com/user-attachments/assets/4e0ef7c9-ae8f-4580-bfcc-8ff9d3acb0f4" />